# Speed up linearization