Hi there @Codsilla and @ngueveu! Following Sandra's awesome presentation at the Julia and Optimization Days, I figured we might be able to gain even more speed compared to the state of the art. Thanks to the tips in my [performance optimization tutorial](https://gdalle.github.io/JuliaOptimizationDays2024-FastJulia/), I achieved the following speed up on a simple test case.

**Current code**

```julia
julia> using BenchmarkTools, LinA

julia> @btime Linearize(abs2, -10, 10, Absolute(0.02));
  4.142 ms (83952 allocations: 1.93 MiB)

julia> @btime Linearize(abs2, -10, 10, Relative(0.02));
  20.275 ms (555988 allocations: 11.34 MiB)
```

**My modifications** (x30 speed up here)

```julia
julia> using BenchmarkTools, LinA

julia> @btime Linearize(abs2, -10, 10, Absolute(0.02));
  129.691 μs (783 allocations: 33.30 KiB)

julia> @btime Linearize(abs2, -10, 10, Relative(0.02));
  543.720 μs (1845 allocations: 99.61 KiB)
```

If you're interested, I can explain to you what I did, and help you improve other aspects of the package.